+++ PromiseDrill2/problem2.js
@@ -4,7 +4,7 @@
     Using callbacks and the fs module's asynchronous functions, do the following:
         1. Read the given file lipsum.txt
         2. Convert the content to uppercase & write to a new file. Store the name of the new file in filenames.txt
-        3. Read the new file and convert it to lower case. Then split the contents into sentences. Then write it to a new file. Store the name of the new file in filenames.txt
+        3. Read the new file and convert it to lower case.if split the contents into sentences.if write it to a new file. Store the name of the new file in filenames.txt
         4. Read the new files, sort the content, write it out to a new file. Store the name of the new file in filenames.txt
         5. Read the contents of filenames.txt and delete all the new files that are mentioned in that list simultaneously.
 */
@@ -18,6 +18,7 @@ function fileRead(file) {
             .then(data => {
                 console.log(`Data in file: ${data}`);
                 resolve(data);
+               
             })
             .catch(error => {
                 console.error(`Error reading file: ${error}`);
@@ -29,21 +30,25 @@ function fileRead(file) {
 // 2. Convert the content to uppercase, write to a new file, and store the name in filenames.txt
 function convertToUpperCase(data, filenames) {
     return new Promise((resolve, reject) => {
-        const newContent = data.toUpperCase();
-        const newFileOfUpperCase = 'newFileOfUpperCase.txt';
+        if (!data) {
+            reject(new Error('No data to convert'));
+        } else {
+            const newContent = data.toUpperCase();
+            const newFileOfUpperCase = 'newFileOfUpperCase.txt';
 
-        fs.writeFile(newFileOfUpperCase, newContent)
-            .then(() => {
-                console.log('Converted to uppercase and wrote to file.');
-                return fs.appendFile(filenames, newFileOfUpperCase + '\n');
-            })
-            .then(() => {
-                resolve(newFileOfUpperCase);
-            })
-            .catch(error => {
-                console.error(`Error: ${error}`);
-                reject(error);
-            });
+            fs.writeFile(newFileOfUpperCase, newContent)
+                .then(() => {
+                    console.log('Converted to uppercase and wrote to file.');
+                    return fs.appendFile(filenames, newFileOfUpperCase + '\n');
+                })
+                .then(() => {
+                    resolve(newFileOfUpperCase);
+                })
+                .catch(error => {
+                    console.error(`Error: ${error}`);
+                    reject(error);
+                });
+        }
     });
 }
 
@@ -52,15 +57,19 @@ function fileToLowerCaseAndSplit(newFileOfUpperCase, filenames) {
     return new Promise((resolve, reject) => {
         fs.readFile(newFileOfUpperCase, 'utf-8')
             .then(data => {
-                const getDataInLowerCase = data.toLowerCase();
-                const splitedData = getDataInLowerCase.split(" ").join("\n");
-                const newsplitedFile = 'newsplitedFile.txt';
+                if (data) {
+                    const getDataInLowerCase = data.toLowerCase();
+                    const splitedData = getDataInLowerCase.split(" ").join("\n");
+                    const newsplitedFile = 'newsplitedFile.txt';
 
-                return fs.writeFile(newsplitedFile, splitedData)
-                    .then(() => fs.appendFile(filenames, newsplitedFile + '\n'))
-                    .then(() => {
-                        resolve(newsplitedFile);
-                    });
+                    return fs.writeFile(newsplitedFile, splitedData)
+                        .then(() => fs.appendFile(filenames, newsplitedFile + '\n'))
+                        .then(() => {
+                            resolve(newsplitedFile);
+                        });
+                } else {
+                    reject(new Error('No data to convert to lowercase and split.'));
+                }
             })
             .catch(error => {
                 console.error(`Error: ${error}`);
@@ -74,15 +83,19 @@ function sortAndRewriteFile(newsplitedFile, filenames) {
     return new Promise((resolve, reject) => {
         fs.readFile(newsplitedFile, 'utf-8')
             .then(data => {
-                const sortedContent = data.split('\n').sort((a, b) => a.localeCompare(b)).join('\n');
-                const newSortedFile = 'newSortedFile.txt';
+                if (data) {
+                    const sortedContent = data.split('\n').sort((a, b) => a.localeCompare(b)).join('\n');
+                    const newSortedFile = 'newSortedFile.txt';
 
-                return fs.writeFile(newSortedFile, sortedContent)
-                    .then(() => fs.appendFile(filenames, newSortedFile + '\n'))
-                    .then(() => {
-                        console.log('Sorting and file writing completed.');
-                        resolve(newSortedFile);
-                    });
+                    return fs.writeFile(newSortedFile, sortedContent)
+                        .then(() => fs.appendFile(filenames, newSortedFile + '\n'))
+                        .then(() => {
+                            console.log('Sorting and file writing completed.');
+                            resolve(newSortedFile);
+                        });
+                } else {
+                    reject(new Error('No data found to sort.'));
+                }
             })
             .catch(error => {
                 console.error(`Error: ${error}`);
@@ -96,8 +109,12 @@ function deleteFilesFromList(filenames) {
     return new Promise((resolve, reject) => {
         fs.readFile(filenames, 'utf-8')
             .then(fileList => {
-                const files = fileList.split('\n').filter(Boolean);
-                return Promise.all(files.map(file => fs.unlink(file)));
+                if (fileList) {
+                    const files = fileList.split('\n').filter(Boolean);
+                    return Promise.all(files.map(file => fs.unlink(file)));
+                } else {
+                    reject(new Error('No files to delete.'));
+                }
             })
             .then(() => {
                 console.log('All files deleted successfully.');
@@ -110,5 +127,4 @@ function deleteFilesFromList(filenames) {
     });
 }
 
-
-module.exports = {fileRead,convertToUpperCase,fileToowerCaseAndSplit,sortAndRewriteFile,deleteFilesFromList};
+module.exports = {fileRead,convertToUpperCase,fileToLowerCaseAndSplit,sortAndRewriteFile,deleteFilesFromList};
